requirements.txt
cryptography==44.0.2
Django==4.2.20
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
